fix: made the RECAPTCHA_SITE_KEY mandatory if RECAPTCHA_SITE_KEY is configured

+++ setup.ts
@@ -214,18 +214,19 @@ async function mongoDB(): Promise<void> {
   }
 }
 
+// Function to ask if the user wants to keep the entered values
+async function askToKeepValues(): Promise<boolean> {
+  const { keepValues } = await inquirer.prompt({
+    type: "confirm",
+    name: "keepValues",
+    message: `Would you like to keep the entered key?`,
+    default: true,
+  });
+  return keepValues;
+}
+
 //Get recaptcha details
 async function recaptcha(): Promise<void> {
-  console.log(
-    "\nPlease visit this URL to set up reCAPTCHA:\n\nhttps://www.google.com/recaptcha/admin/create"
-  );
-  console.log(
-    '\nSelect reCAPTCHA v2 and the "I`m not a robot" checkbox option'
-  );
-  console.log(
-    '\nAdd "localhost" in domains and accept the terms, then press submit'
-  );
-
   const { recaptchaSecretKey } = await inquirer.prompt([
     {
       type: "input",
@@ -239,28 +240,32 @@ async function recaptcha(): Promise<void> {
       },
     },
   ]);
-  const config = dotenv.parse(fs.readFileSync(".env"));
-  config.RECAPTCHA_SECRET_KEY = recaptchaSecretKey;
-  fs.writeFileSync(".env", "");
-  for (const key in config) {
-    fs.appendFileSync(".env", `${key}=${config[key]}\n`);
+
+  const shouldKeepDetails = await askToKeepValues();
+
+  if (shouldKeepDetails) {
+    const config = dotenv.parse(fs.readFileSync(".env"));
+    config.RECAPTCHA_SECRET_KEY = recaptchaSecretKey;
+    fs.writeFileSync(".env", "");
+    for (const key in config) {
+      fs.appendFileSync(".env", `${key}=${config[key]}\n`);
+    }
+  } else {
+    await recaptcha();
   }
 }
-async function recaptchaSiteKey(): Promise<void> {
-  console.log(
-    "\nPlease visit this URL to set up reCAPTCHA:\n\nhttps://www.google.com/recaptcha/admin/create"
-  );
-  console.log(
-    '\nSelect reCAPTCHA v2 and the "I`m not a robot" checkbox option'
-  );
-  console.log(
-    '\nAdd "localhost" in domains and accept the terms, then press submit'
-  );
 
-  const { recaptchaSiteKey } = await inquirer.prompt([
+async function recaptchaSiteKey(): Promise<void> {
+  if (process.env.RECAPTCHA_SITE_KEY) {
+    console.log(
+      `\nreCAPTCHA site key already exists with the value ${process.env.RECAPTCHA_SITE_KEY}`
+    );
+  }
+
+  const { recaptchaSiteKeyInp } = await inquirer.prompt([
     {
       type: "input",
-      name: "recaptchaSiteKey",
+      name: "recaptchaSiteKeyInp",
       message: "Enter your reCAPTCHA site key:",
       validate: async (input: string): Promise<boolean | string> => {
         if (validateRecaptcha(input)) {
@@ -270,11 +275,18 @@ async function recaptchaSiteKey(): Promise<void> {
       },
     },
   ]);
-  const config = dotenv.parse(fs.readFileSync(".env"));
-  config.RECAPTCHA_SITE_KEY = recaptchaSiteKey;
-  fs.writeFileSync(".env", "");
-  for (const key in config) {
-    fs.appendFileSync(".env", `${key}=${config[key]}\n`);
+
+  const shouldKeepDetails = await askToKeepValues();
+
+  if (shouldKeepDetails) {
+    const config = dotenv.parse(fs.readFileSync(".env"));
+    config.RECAPTCHA_SITE_KEY = recaptchaSiteKeyInp;
+    fs.writeFileSync(".env", "");
+    for (const key in config) {
+      fs.appendFileSync(".env", `${key}=${config[key]}\n`);
+    }
+  } else {
+    await recaptchaSiteKey();
   }
 }
 
@@ -425,7 +437,7 @@ async function main(): Promise<void> {
     type: "confirm",
     name: "shouldGenerateAccessToken",
     message: "Would you like to generate a new access token secret?",
-    default: true,
+    default: process.env.ACCESS_TOKEN_SECRET ? false : true,
   });
 
   if (shouldGenerateAccessToken) {
@@ -441,7 +453,7 @@ async function main(): Promise<void> {
     type: "confirm",
     name: "shouldGenerateRefreshToken",
     message: "Would you like to generate a new refresh token secret?",
-    default: true,
+    default: process.env.REFRESH_TOKEN_SECRET ? false : true,
   });
 
   if (shouldGenerateRefreshToken) {
@@ -467,7 +479,7 @@ async function main(): Promise<void> {
       type: "confirm",
       name: "shouldSetRedis",
       message: "Would you like to set up a Redis URL?",
-      default: true,
+      default: process.env.REDIS_URL ? false : true,
     });
     if (shouldSetRedis) {
       await redisConfiguration();
@@ -483,7 +495,7 @@ async function main(): Promise<void> {
       type: "confirm",
       name: "shouldSetMongoDb",
       message: "Would you like to set up a MongoDB URL?",
-      default: true,
+      default: process.env.MONGO_DB_URL ? false : true,
     });
 
     if (shouldSetMongoDb) {
@@ -499,21 +511,11 @@ async function main(): Promise<void> {
     type: "confirm",
     name: "shouldSetRecaptcha",
     message: "Would you like to set up a reCAPTCHA secret key?",
-    default: true,
+    default: process.env.RECAPTCHA_SECRET_KEY ? false : true,
   });
 
   if (shouldSetRecaptcha) {
     await recaptcha();
-  }
-
-  const { shouldSetRecaptchaSiteKey } = await inquirer.prompt({
-    type: "confirm",
-    name: "shouldSetRecaptchaSiteKey",
-    message: "Would you like to set up a reCAPTCHA site key?",
-    default: true,
-  });
-
-  if (shouldSetRecaptchaSiteKey) {
     await recaptchaSiteKey();
   }
 
@@ -527,6 +529,7 @@ async function main(): Promise<void> {
       type: "confirm",
       name: "shouldSetMail",
       message: "Would you like to setup the mail username and password?",
+      default: process.env.MAIL_USERNAME ? false : true,
     },
   ]);
   if (shouldSetMail) {
@@ -544,7 +547,7 @@ async function main(): Promise<void> {
       type: "confirm",
       name: "shouldSetSuperUserEmail",
       message: "Would you like to setup a Super Admin email of last resort?",
-      default: true,
+      default: process.env.LAST_RESORT_SUPERADMIN_EMAIL ? false : true,
     },
   ]);
   if (shouldSetSuperUserEmail) {
